Moving route to v2 api
