+++ pkg/handler/v1/provider/openstack.go
@@ -21,7 +21,6 @@ import (
 	"crypto/x509"
 	"fmt"
 	"net/http"
-	"strconv"
 	"strings"
 
 	"github.com/go-kit/kit/endpoint"
@@ -34,7 +33,7 @@ import (
 	utilerrors "k8c.io/kubermatic/v2/pkg/util/errors"
 )
 
-func getAuthInfo(ctx context.Context, req OpenstackReq, userInfoGetter provider.UserInfoGetter, presetProvider provider.PresetProvider) (*provider.UserInfo, *resources.OpenstackCredentials, error) {
+func GetOpenstackAuthInfo(ctx context.Context, req OpenstackReq, userInfoGetter provider.UserInfoGetter, presetProvider provider.PresetProvider) (*provider.UserInfo, *resources.OpenstackCredentials, error) {
 	var cred *resources.OpenstackCredentials
 	userInfo, err := userInfoGetter(ctx, "")
 	if err != nil {
@@ -79,7 +78,7 @@ func OpenstackSizeEndpoint(seedsGetter provider.SeedsGetter, presetProvider prov
 		if !ok {
 			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackReq, got %T", request)
 		}
-		userInfo, cred, err := getAuthInfo(ctx, req, userInfoGetter, presetProvider)
+		userInfo, cred, err := GetOpenstackAuthInfo(ctx, req, userInfoGetter, presetProvider)
 		if err != nil {
 			return nil, err
 		}
@@ -126,7 +125,7 @@ func OpenstackTenantEndpoint(seedsGetter provider.SeedsGetter, presetProvider pr
 			OIDCAuthentication:          reqTenant.OIDCAuthentication,
 		}
 
-		userInfo, cred, err := getAuthInfo(ctx, req, userInfoGetter, presetProvider)
+		userInfo, cred, err := GetOpenstackAuthInfo(ctx, req, userInfoGetter, presetProvider)
 		if err != nil {
 			return nil, err
 		}
@@ -152,7 +151,7 @@ func OpenstackNetworkEndpoint(seedsGetter provider.SeedsGetter, presetProvider p
 		if !ok {
 			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackReq, got %T", request)
 		}
-		userInfo, cred, err := getAuthInfo(ctx, req, userInfoGetter, presetProvider)
+		userInfo, cred, err := GetOpenstackAuthInfo(ctx, req, userInfoGetter, presetProvider)
 		if err != nil {
 			return nil, err
 		}
@@ -177,7 +176,7 @@ func OpenstackSecurityGroupEndpoint(seedsGetter provider.SeedsGetter, presetProv
 		if !ok {
 			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackReq, got %T", request)
 		}
-		userInfo, cred, err := getAuthInfo(ctx, req, userInfoGetter, presetProvider)
+		userInfo, cred, err := GetOpenstackAuthInfo(ctx, req, userInfoGetter, presetProvider)
 		if err != nil {
 			return nil, err
 		}
@@ -202,7 +201,7 @@ func OpenstackSubnetsEndpoint(seedsGetter provider.SeedsGetter, presetProvider p
 		if !ok {
 			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackSubnetReq, got %T", request)
 		}
-		userInfo, cred, err := getAuthInfo(ctx, req.OpenstackReq, userInfoGetter, presetProvider)
+		userInfo, cred, err := GetOpenstackAuthInfo(ctx, req.OpenstackReq, userInfoGetter, presetProvider)
 		if err != nil {
 			return nil, err
 		}
@@ -227,7 +226,7 @@ func OpenstackAvailabilityZoneEndpoint(seedsGetter provider.SeedsGetter, presetP
 		if !ok {
 			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackReq, got %T", request)
 		}
-		userInfo, cred, err := getAuthInfo(ctx, req, userInfoGetter, presetProvider)
+		userInfo, cred, err := GetOpenstackAuthInfo(ctx, req, userInfoGetter, presetProvider)
 		if err != nil {
 			return nil, err
 		}
@@ -247,21 +246,6 @@ func OpenstackAvailabilityZoneWithClusterCredentialsEndpoint(projectProvider pro
 		req := request.(OpenstackNoCredentialsReq)
 		return providercommon.OpenstackAvailabilityZoneWithClusterCredentialsEndpoint(ctx, userInfoGetter, projectProvider,
 			privilegedProjectProvider, seedsGetter, req.ProjectID, req.ClusterID, caBundle)
-	}
-}
-
-func OpenstackSubnetPoolEndpoint(seedsGetter provider.SeedsGetter, presetProvider provider.PresetProvider,
-	userInfoGetter provider.UserInfoGetter, caBundle *x509.CertPool) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req, ok := request.(OpenstackSubnetPoolReq)
-		if !ok {
-			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackReq, got %T", request)
-		}
-		userInfo, cred, err := getAuthInfo(ctx, req.OpenstackReq, userInfoGetter, presetProvider)
-		if err != nil {
-			return nil, err
-		}
-		return providercommon.GetOpenstackSubnetPools(ctx, userInfo, seedsGetter, cred, req.DatacenterName, req.IPVersion, caBundle)
 	}
 }
 
@@ -349,34 +333,6 @@ func DecodeOpenstackSubnetNoCredentialsReq(c context.Context, r *http.Request) (
 	if req.NetworkID == "" {
 		return nil, fmt.Errorf("get openstack subnets needs a parameter 'network_id'")
 	}
-	return req, nil
-}
-
-// OpenstackSubnetPoolReq represent a request for openstack subnet pools
-// swagger:parameters listOpenstackSubnetPools
-type OpenstackSubnetPoolReq struct {
-	OpenstackReq
-	// in: query
-	IPVersion int `json:"ip_version,omitempty"`
-}
-
-func DecodeOpenstackSubnetPoolReq(_ context.Context, r *http.Request) (interface{}, error) {
-	var req OpenstackSubnetPoolReq
-
-	openstackReq, err := DecodeOpenstackReq(context.Background(), r)
-	if err != nil {
-		return nil, err
-	}
-	req.OpenstackReq = openstackReq.(OpenstackReq)
-
-	ipVersion := r.URL.Query().Get("ip_version")
-	if ipVersion != "" {
-		req.IPVersion, err = strconv.Atoi(ipVersion)
-		if err != nil || (req.IPVersion != 4 && req.IPVersion != 6) {
-			return nil, utilerrors.NewBadRequest("invalid value for `ip_version` (should be 4 or 6)")
-		}
-	}
-
 	return req, nil
 }
