+++ pkg/handler/v2/provider/openstack.go
@@ -21,14 +21,17 @@ import (
 	"crypto/x509"
 	"fmt"
 	"net/http"
+	"strconv"
 
 	"github.com/go-kit/kit/endpoint"
 
 	apiv1 "k8c.io/kubermatic/v2/pkg/api/v1"
 	providercommon "k8c.io/kubermatic/v2/pkg/handler/common/provider"
 	"k8c.io/kubermatic/v2/pkg/handler/v1/common"
+	providerv1 "k8c.io/kubermatic/v2/pkg/handler/v1/provider"
 	"k8c.io/kubermatic/v2/pkg/handler/v2/cluster"
 	"k8c.io/kubermatic/v2/pkg/provider"
+	utilerrors "k8c.io/kubermatic/v2/pkg/util/errors"
 )
 
 func OpenstackSizeWithClusterCredentialsEndpoint(projectProvider provider.ProjectProvider,
@@ -91,6 +94,21 @@ func OpenstackAvailabilityZoneWithClusterCredentialsEndpoint(projectProvider pro
 	}
 }
 
+func OpenstackSubnetPoolEndpoint(seedsGetter provider.SeedsGetter, presetProvider provider.PresetProvider,
+	userInfoGetter provider.UserInfoGetter, caBundle *x509.CertPool) endpoint.Endpoint {
+	return func(ctx context.Context, request interface{}) (interface{}, error) {
+		req, ok := request.(OpenstackSubnetPoolReq)
+		if !ok {
+			return nil, fmt.Errorf("incorrect type of request, expected = OpenstackReq, got %T", request)
+		}
+		userInfo, cred, err := providerv1.GetOpenstackAuthInfo(ctx, req.OpenstackReq, userInfoGetter, presetProvider)
+		if err != nil {
+			return nil, err
+		}
+		return providercommon.GetOpenstackSubnetPools(ctx, userInfo, seedsGetter, cred, req.DatacenterName, req.IPVersion, caBundle)
+	}
+}
+
 // openstackNoCredentialsReq represent a request for openstack
 // swagger:parameters listOpenstackSizesNoCredentialsV2 listOpenstackTenantsNoCredentialsV2 listOpenstackNetworksNoCredentialsV2 listOpenstackSecurityGroupsNoCredentialsV2 listOpenstackAvailabilityZonesNoCredentialsV2
 type openstackNoCredentialsReq struct {
@@ -147,6 +165,34 @@ func DecodeOpenstackSubnetNoCredentialsReq(c context.Context, r *http.Request) (
 	req.NetworkID = r.URL.Query().Get("network_id")
 	if req.NetworkID == "" {
 		return nil, fmt.Errorf("get openstack subnets needs a parameter 'network_id'")
+	}
+
+	return req, nil
+}
+
+// OpenstackSubnetPoolReq represent a request for openstack subnet pools
+// swagger:parameters listOpenstackSubnetPools
+type OpenstackSubnetPoolReq struct {
+	providerv1.OpenstackReq
+	// in: query
+	IPVersion int `json:"ip_version,omitempty"`
+}
+
+func DecodeOpenstackSubnetPoolReq(_ context.Context, r *http.Request) (interface{}, error) {
+	var req OpenstackSubnetPoolReq
+
+	openstackReq, err := providerv1.DecodeOpenstackReq(context.Background(), r)
+	if err != nil {
+		return nil, err
+	}
+	req.OpenstackReq = openstackReq.(providerv1.OpenstackReq)
+
+	ipVersion := r.URL.Query().Get("ip_version")
+	if ipVersion != "" {
+		req.IPVersion, err = strconv.Atoi(ipVersion)
+		if err != nil || (req.IPVersion != 4 && req.IPVersion != 6) {
+			return nil, utilerrors.NewBadRequest("invalid value for `ip_version` (should be 4 or 6)")
+		}
 	}
 
 	return req, nil
